requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Speech processing
speechbrain>=0.5.16
openai-whisper>=20230314
librosa>=0.10.0
soundfile>=0.12.0

# Audio quality metrics
pesq>=0.0.4
pystoi>=0.3.3

# Data processing
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

